Removed Deprecated UserEntity Interface - Moved ExtraUser from org.mojodojocasahouse.model.impl to org.mojodojocasahouse.model
