--- src/main/java/org/mojodojocasahouse/extra/model/UserEntity.java
@@ -1,4 +1,0 @@
-package org.mojodojocasahouse.extra.model;
-
-public interface UserEntity {
-}
