+++ src/main/java/org/mojodojocasahouse/extra/repository/ExtraUserRepository.java
@@ -1,7 +1,7 @@
 package org.mojodojocasahouse.extra.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
-import org.mojodojocasahouse.extra.model.impl.ExtraUser;
+import org.mojodojocasahouse.extra.model.ExtraUser;
 import org.springframework.stereotype.Repository;
 
 import java.util.Optional;
