+++ src/test/java/org/mojodojocasahouse/extra/repository/ExtraUserRepositoryTest.java
@@ -2,17 +2,12 @@ package org.mojodojocasahouse.extra.repository;
 
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.Test;
-import org.junit.runner.RunWith;
-import org.mojodojocasahouse.extra.exception.ExistingUserEmailException;
-import org.mojodojocasahouse.extra.model.impl.ExtraUser;
+import org.mojodojocasahouse.extra.model.ExtraUser;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.autoconfigure.orm.jpa.DataJpaTest;
 import org.springframework.dao.DataIntegrityViolationException;
-import org.springframework.test.context.junit4.SpringRunner;
 
 import java.util.Optional;
-
-import static org.junit.jupiter.api.Assertions.*;
 
 
 @DataJpaTest
